tests/FACE/Unit does not conform to guidelines
Problem
-------

Unit tests under tests/FACE/Unit do not conform to
/docs/internal/unit_tests.rst.

Solution
--------

Move the unit tests according to the guidelines.
