--- tests/FACE/Unit/.gitignore
@@ -1,7 +1,0 @@
-/QosSettingsTest
-/PoolTest
-/PoolAllocationTest
-/PoolAllocatorTest
-/SequencesTest
-/MemoryPoolTest
-/FreeIndexTest
