--- tests/FACE/Unit/FaceUnitTests.mpc
@@ -1,23 +1,0 @@
-project(*QosSettings) : dcps, opendds_face, dcps_test {
-  Source_Files {
-    QosSettingsTest.cpp
-  }
-}
-
-project(*MemoryPool) : dcps, opendds_face, dcps_test {
-  Source_Files {
-    MemoryPoolTest.cpp
-  }
-}
-
-project(*FreeIndex) : dcps, opendds_face, dcps_test {
-  Source_Files {
-    FreeIndexTest.cpp
-  }
-}
-
-project(*Sequences): dcps, opendds_face, dcps_test {
-  Source_Files {
-    SequencesTest.cpp
-  }
-}
