+++ tests/unit-tests/FACE/UnitTests.mpc
@@ -1,0 +1,6 @@
+project(*Sequences): dcps, opendds_face, dcps_test {
+  exename = *
+  Source_Files {
+    Sequence.cpp
+  }
+}
